added new audio engine with alsa

+++ include/engine/wav-reader.h
@@ -1,6 +1,9 @@
 #ifndef WAV_READER_H_
 #define WAV_READER_H_
 
-float *WavReader_getDataFromWavFile(char *, int *);
+#include "stddef.h"
+
+//get data and number of pcm frames
+int16_t *WavReader_getDataFromWavFile(char *, int *);
 
 #endif
